+++ front/scripts/main/app.ts
@@ -1,6 +1,7 @@
 /// <reference path="../../../typings/jquery/jquery.d.ts" />
 /// <reference path="../../../typings/ember/ember.d.ts" />
 /// <reference path="../../../typings/i18next/i18next.d.ts" />
+/// <reference path="../../../typings/jquery.cookie/jquery.cookie.d.ts" />
 /// <reference path="../baseline/mercury.ts" />
 /// <reference path="../mercury/modules/Ads.ts" />
 /// <reference path="../mercury/modules/Trackers/UniversalAnalytics.ts" />
@@ -19,6 +20,7 @@ interface Window {
 declare var i18n: I18nextStatic;
 declare var EmPerfSender: any;
 declare var optimizely: any;
+declare var FastClick: any;
 
 var App: any = Em.Application.create({
 	// We specify a rootElement, otherwise Ember appends to the <body> element and Google PageSpeed thinks we are
@@ -70,6 +72,10 @@ App.initializer({
 			resStore: loadedTranslations,
 			useLocalStorage: false
 		});
+
+		// FastClick disables the 300ms delay on iOS and some Android devices. It also uses clicks so that
+		// elements have access to :hover state
+		FastClick.attach(document.body);
 	}
 });
 
@@ -148,6 +154,23 @@ App.initializer({
 		dimensions = Mercury.Utils.VariantTesting.integrateOptimizelyWithUA(dimensions);
 
 		UA.setDimensions(dimensions);
+	}
+});
+
+/**
+ * A "Geo" cookie is set by Fastly on every request.
+ * If you run mercury app on your laptop, the cookie won't be automatically present.
+ */
+App.initializer({
+	name: 'geo',
+	after: 'setupTracking',
+	initialize(container: any, application: typeof App): void {
+		var geoCookie = $.cookie('Geo');
+		if (geoCookie) {
+			M.prop('geo', JSON.parse(geoCookie));
+		} else {
+			Ember.debug('Geo cookie is not set');
+		}
 	}
 });