+++ front/scripts/main/models/CuratedContentModel.ts
@@ -13,6 +13,7 @@ interface CuratedContentItem {
 	type: string;
 	url?: string;
 	categoryName?: string;
+	ns?: number;
 }
 
 App.CuratedContentModel = Em.Object.extend({
@@ -20,8 +21,10 @@ App.CuratedContentModel = Em.Object.extend({
 		return new Em.RSVP.Promise((resolve: Function, reject: Function): void => {
 			var url = App.get('apiBase');
 			url += (sectionType === 'section') ?
-				'/curatedContent/' + sectionName :
-				'/category/' + sectionName;
+				//We don't need to wrap it into Try/Catch statement
+				//See: https://github.com/Wikia/mercury/pull/946#issuecomment-113501147
+				'/curatedContent/' + encodeURIComponent(sectionName) :
+				'/category/' + encodeURIComponent(sectionName);
 
 			Em.$.ajax({
 				url: url,
@@ -64,6 +67,11 @@ App.CuratedContentModel = Em.Object.extend({
 				imageUrl: rawData.thumbnail,
 				type: rawData.type,
 				url: rawData.url
+			};
+
+			// ArticlesApi doesn't return type for blog posts so we need to look at the namespace
+			if (Em.isEmpty(rawData.type) && rawData.ns === 500) {
+				item.type = 'blog';
 			}
 		}