+++ server/lib/AuthUtils.ts
@@ -38,3 +38,9 @@ export function getCacheBusterUrl(redirect: string): string {
 
 	return url.format(cacheBustedUrlObj);
 }
+
+export function getRedirectUrlWithQueryString(route: string, request: Hapi.Request) {
+	var redirectUrl = request.url;
+	redirectUrl.pathname = route;
+	return redirectUrl.format();
+}